+++ ChangeLog.md
@@ -1,4 +1,11 @@
-﻿## 07/27/2019 V8.8.1
+﻿## 08/20/2019 V8.9.0
+
+- Add `Console.Erase`
+- Add `Console.GetComponent` and `Console.GetComponents` to find components added to a console.
+- Add `CellSurface.CreateLine` to create an array of int's for a line.
+- Instructions supports a `RemoveOnFinished` property now.
+
+## 07/27/2019 V8.8.1
 
 - Fix docs